Initial push
This is the initial push for my nvim config files.
I decide to move it from my dotfiles and create a separate repo, because
I don't want to use stow anymore, and I start to use splitted files to
manage my plugins, and it would be hard to keep tracking everything in
ansible. So, now the idea is that I can from my dotfiles have a ansible
task that will just clone this repo in $HOME/.config/nvim and done.